+++ src/TorchSharp/Optimizers/Adamax.cs
@@ -148,7 +148,7 @@ namespace TorchSharp
 
                     foreach (var param in group.Parameters) {
 
-                        var grad = param.grad();
+                        var grad = param.grad;
 
                         if (grad is null) continue;