+++ RELEASENOTES.md
@@ -8,6 +8,8 @@ __Breaking Changes__:
 
 __API Changes__:
 
+- #1291 `Tensor.grad()` and `Tensor.set_grad()` have been replaced by a new property `Tensor.grad`.
+
 __Bug Fixes__: